+++ R/RcppExports.R
@@ -21,3 +21,11 @@ dmvtCpp <- function(X_, mu_, sigma_, df_, log_, ncores_, isChol_) {
     .Call('_pRoloc_dmvtCpp', PACKAGE = 'pRoloc', X_, mu_, sigma_, df_, log_, ncores_, isChol_)
 }
 
+dmvt_log_batch_consistent <- function(X, mu, Sigma, df) {
+    .Call('_pRoloc_dmvt_log_batch_consistent', PACKAGE = 'pRoloc', X, mu, Sigma, df)
+}
+
+dmvt_log_row_multi_consistent <- function(x, MU, SIGMA_list, df) {
+    .Call('_pRoloc_dmvt_log_row_multi_consistent', PACKAGE = 'pRoloc', x, MU, SIGMA_list, df)
+}
+